refs 125405: move external user group reelated files to features
