+++ apps/app/src/features/external-user-group/client/components/ExternalUserGroup/LdapGroupManagement.tsx
