+++ apps/app/src/features/external-user-group/client/stores/external-user-group.ts
